number type changes

+++ Dev_12-20/sport.py
@@ -1,4 +1,4 @@
-age=float(input('请输入年龄: '))
+age=int(input('请输入年龄: '))
 sex=input("请输入性别 为male / female: ")
 if(sex == "male"):
     n = int(220)
@@ -10,6 +10,6 @@ else:
         exit()
 HRrest = int(input("请输入安静心率: "))
 
-low=(n-age-HRrest)*0.6+HRrest
-high=(n-age-HRrest)*0.8+HRrest
+low=float((n-age-HRrest)*0.6+HRrest)
+high=float((n-age-HRrest)*0.8+HRrest)
 print("您的最佳运动心率为:",low,"~",high)
